+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### unreleased
+
+* Adds support to rails test environment
+
 ### v0.1.0
 
 * Initial release.